+++ mchplnet/interfaces/lin.py
@@ -1,4 +1,4 @@
-from pylnet.interfaces.abstract_interface import InterfaceABC
+from mchplnet.interfaces.abstract_interface import InterfaceABC
 
 
 class LNetLin(InterfaceABC):
