+++ mchplnet/services/device_info.py
