+++ mchplnet/services/frame_putram.py
@@ -1,6 +1,6 @@
 import logging
 
-from pylnet.lnetframe import LNetFrame
+from mchplnet.lnetframe import LNetFrame
 
 
 class FramePutRam(LNetFrame):
